Retail_Inventory_management_tshirts
This is an end to end LLM project based on Google Palm and Langchain. We are building a system that can talk to MySQL database. User asks questions in a natural language and the system generates answers by converting those questions to an SQL query and then executing that query on MySQL database. NK Tees is a T-shirt store where they maintain their inventory, sales and discounts data in MySQL database.